ISP-35 Procced seller verification before submit product changes.

+++ routes/products/actions.js
@@ -61,7 +61,7 @@ router.get('/view-all/:uuidTag', async function(req, res, next) {
 
 
 
-router.post('/interact', async function(req, res, next) {
+router.post('/changes', async function(req, res, next) {
     try {
         function isValid(body) {
             if(body === undefined) throw new Error("Unexpected empty body")
@@ -72,7 +72,6 @@ router.post('/interact', async function(req, res, next) {
         }
         isValid(req.body)
     } catch(e) {
-        console.log(req.body)
         res.status(400).json({"error_message" : "Unexpected request body"})
         return
     }
@@ -87,9 +86,21 @@ router.post('/interact', async function(req, res, next) {
         return
     }
 
-    const productUUID = requestBody['product_uuid']
+    if(isValidSession.rows[0]['seller_token_expiration_date'] < currentDate){
+        const deleteExpiredToken = `UPDATE ${database.Tables.sellers} SET
+                                        seller_authentication_token = null,
+                                        seller_token_in_use = false,
+                                        seller_token_creation_date = null,
+                                        seller_token_expiration_date = null
+                                    WHERE seller_authentication_token = $1;`
+        await database.query(deleteExpiredToken, requestBody['session_token'])
+        res.status(400).json({"error_message" : "Session expired please login again."})
+        return
+    }
 
+    const productUUID = requestBody['product_uuid']
     const sellerUUID = isValidSession.rows[0]['seller_unique_register_id']
+
     switch(requestBody.action) {
         case "add" :
             const addStmt = `INSERT INTO ${database.Tables.products} (
@@ -99,8 +110,9 @@ router.post('/interact', async function(req, res, next) {
                                  product_quantity,
                                  product_company_owner_uuid,
                                  product_description,
+                                 product_image_data,
                                  product_register_date,
-                                 product_status)  VALUES ($1, $2, $3, $4, $5, $6, $7, $8) RETURNING true;`
+                                 product_status)  VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9) RETURNING true;`
             const addParams = [
                 uuid.v4(),
                 requestBody.product.name,
@@ -108,6 +120,7 @@ router.post('/interact', async function(req, res, next) {
                 requestBody.product.quantity,
                 sellerUUID,
                 requestBody.product.description,
+                requestBody.product.image,
                 'NOW()',
                 requestBody.product.quantity <= 0 ? 0 : 1
             ]
@@ -122,17 +135,21 @@ router.post('/interact', async function(req, res, next) {
 
         case "edit" :
             const editStmt = `UPDATE ${database.Tables.products} SET
-                                product_name = $1,
-                                product_price = $2,
-                                product_quantity = $3,
-                                product_description = $4
-                               WHERE product_unique_register_id = $5 RETURNING true;`
+                                  product_name = $1,
+                                  product_price = $2,
+                                  product_quantity = $3,
+                                  product_description = $4,
+                                  product_image_data = $5
+                              WHERE (product_unique_register_id = $6 
+                              AND product_company_owner_uuid = $7) RETURNING true;`
             const editParams = [
                 requestBody.product.name,
                 requestBody.product.price,
                 requestBody.product.quantity,
                 requestBody.product.description,
-                productUUID
+                requestBody.product.image,
+                productUUID,
+                sellerUUID
             ]
             const editResponse = await database.query(editStmt, editParams)
             if(editResponse.rows.length > 0){
@@ -144,8 +161,10 @@ router.post('/interact', async function(req, res, next) {
 
         case "remove" :
             const removeStmt = `DELETE FROM ${database.Tables.products}
-                                WHERE product_unique_register_id = $1 RETURNING true;`
-            const removeResponse = await database.query(removeStmt, [productUUID])
+                                WHERE product_unique_register_id = $1 
+                                AND product_company_owner_uuid = $2
+                                RETURNING true;`
+            const removeResponse = await database.query(removeStmt, [productUUID, sellerUUID])
             if(removeResponse.rows.length > 0){
                 res.status(201).end()
             } else {
@@ -154,7 +173,7 @@ router.post('/interact', async function(req, res, next) {
             break
 
         default:
-            res.status(400).json({"error_message" : "Invalid interaction provided"})
+            res.status(400).json({"error_message" : "Invalid action provided"})
             break
     }
 })
